trying int instead of ssize_t as return value for device_write

+++ iodriver.c
@@ -14,7 +14,7 @@ static char data[MAX];
 static int device_open(struct inode *,struct file *);
 static int device_release(struct inode *,struct file *);
 static ssize_t device_read(struct file *,char *,size_t ,loff_t *);
-static ssize_t device_write(struct file *,char *,size_t,loff_t *);
+static int device_write(struct file *,char *,size_t,loff_t *);
 
 static struct file_operations fops = 
 {
@@ -78,7 +78,7 @@ static ssize_t device_read(struct file *	filp,char *buffer,size_t size, loff_t *
 	return bytesRead;
 }
 
-static ssize_t device_write(struct file *filp,char *buffer,size_t size, loff_t *offset)
+static int device_write(struct file *filp,char *buffer,size_t size, loff_t *offset)
 {
 	char *ptrData = data;
 	int bytesWritten = 0;
